@@ -1,6 +1,6 @@
 package com.yveschiong.easycalendar.sample.activities;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 public abstract class BaseActivity extends AppCompatActivity {
 
