@@ -1,9 +1,10 @@
 package com.yveschiong.easycalendar.sample.activities;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.yveschiong.easycalendar.sample.App;
 import com.yveschiong.easycalendar.sample.R;
