@@ -2,7 +2,8 @@ package com.yveschiong.easycalendar.sample.helpers;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.yveschiong.easycalendar.sample.activities.BaseActivity;
 
