@@ -1,10 +1,11 @@
 package com.yveschiong.easycalendar.sample.adapters;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.annotation.NonNull;
 
 import com.yveschiong.easycalendar.sample.R;
 import com.yveschiong.easycalendar.sample.presenters.ActivityListPresenter;
